Moved Types
- Moved Types to public folder
